ai_inference/requirements.txt
# Member 3 – AI Skin Disease Classification
tensorflow>=2.15.0,<3
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
scikit-learn>=1.2.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
# For downloading from Kaggle / multiple sources
kaggle>=1.5.0
